Table moved to sepatare component with default configurations
