+++ src/frontend/src/pages/ListsPage.js
@@ -3,47 +3,10 @@ import { useListsState, ListsStateActions } from './../State/ListsState'
 import DateTimeFormatter from './../Common/DateTimeFormatter'
 import useListService from './../Services/ListService'
 import { useModalState, ModalStateAction, ModalTypes } from './../State/ModalStateProvider'
+import Table from './../components/Table'
 
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { withStyles } from '@material-ui/core/styles';
-
-import { forwardRef } from 'react';
-import MaterialTable from 'material-table';
-import AddBox from '@material-ui/icons/AddBox';
-import ArrowDownward from '@material-ui/icons/ArrowDownward';
-import Check from '@material-ui/icons/Check';
-import ChevronLeft from '@material-ui/icons/ChevronLeft';
-import ChevronRight from '@material-ui/icons/ChevronRight';
-import Clear from '@material-ui/icons/Clear';
-import DeleteOutline from '@material-ui/icons/DeleteOutline';
-import Edit from '@material-ui/icons/Edit';
-import FilterList from '@material-ui/icons/FilterList';
-import FirstPage from '@material-ui/icons/FirstPage';
-import LastPage from '@material-ui/icons/LastPage';
-import Remove from '@material-ui/icons/Remove';
-import SaveAlt from '@material-ui/icons/SaveAlt';
-import Search from '@material-ui/icons/Search';
-import ViewColumn from '@material-ui/icons/ViewColumn';
-
-const tableIcons = {
-    Add: forwardRef((props, ref) => <AddBox {...props} ref={ref} />),
-    Check: forwardRef((props, ref) => <Check {...props} ref={ref} />),
-    Clear: forwardRef((props, ref) => <Clear {...props} ref={ref} />),
-    Delete: forwardRef((props, ref) => <DeleteOutline {...props} ref={ref} />),
-    DetailPanel: forwardRef((props, ref) => <ChevronRight {...props} ref={ref} />),
-    Edit: forwardRef((props, ref) => <Edit {...props} ref={ref} />),
-    Export: forwardRef((props, ref) => <SaveAlt {...props} ref={ref} />),
-    Filter: forwardRef((props, ref) => <FilterList {...props} ref={ref} />),
-    FirstPage: forwardRef((props, ref) => <FirstPage {...props} ref={ref} />),
-    LastPage: forwardRef((props, ref) => <LastPage {...props} ref={ref} />),
-    NextPage: forwardRef((props, ref) => <ChevronRight {...props} ref={ref} />),
-    PreviousPage: forwardRef((props, ref) => <ChevronLeft {...props} ref={ref} />),
-    ResetSearch: forwardRef((props, ref) => <Clear {...props} ref={ref} />),
-    Search: forwardRef((props, ref) => <Search {...props} ref={ref} />),
-    SortArrow: forwardRef((props, ref) => <ArrowDownward {...props} ref={ref} />),
-    ThirdStateCheck: forwardRef((props, ref) => <Remove {...props} ref={ref} />),
-    ViewColumn: forwardRef((props, ref) => <ViewColumn {...props} ref={ref} />)
-  };
 
 const styles = {
     container: {
@@ -77,7 +40,7 @@ const ListsPage = (props) => {
         <div className={props.classes.container}>
             {isLoading 
             ? <CircularProgress />
-            : <MaterialTable
+            : <Table
                 title={`You have ${listState.lists.length} lists`}
                 columns={[
                     {title: "Name", field: "name"},
@@ -89,13 +52,7 @@ const ListsPage = (props) => {
                     //TODO => {title: "Shared", field: "isShared", render: isShared => isShared ? "Yes" : "No"},
                 ]}
                 onRowClick={((evt, selectedRow) => console.log(selectedRow.id))}
-                options={{
-                    pageSize: 5,
-                    pageSizeOptions: [],
-                    actionsColumnIndex: -1
-                }}
-                data={listState.lists} 
-                icons={tableIcons} 
+                data={listState.lists}
                 actions={[
                     {
                         icon: 'edit',
